Update patch set 1
Patch Set 1:

I think this was the wrong way of solving the problem, because it makes the message "found #%s tests at %s" that we log in main meaningless (because the plugin path is not defined at that point).  A better way of handling the issue w/r/t to my original commit would have been to -1 the commit and request a change.  One way of making sure this doesn't happen in the future is to require reviews from two different people before merging, although this may not make sense with the current number of developers.

Patch-set: 1
Reviewer: Lars Kellogg-Stedman <1001152@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0