Update patch set 1
Patch Set 1:

We are importing the functions from citellus into Magui.
We need to ensure we don't remove the functions that are used in Magui. Or we can move them out into library.

Also during previous updates we have merged broken the Magui. To fix this we are looking for a way to have CI voting on the commits like it did in Github PRs

Patch-set: 1
Reviewer: Robin Černín <1011950@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0