Update patch set 2
Patch Set 2:

Yes,
It's a template file that can and should be generated with ./extractpot.sh to ensure it's using latest version.

If for each plugin we update we also update citellus.pot there are higher chances of getting collisions when merging reviews as both of reviews will target citellus.pot and fail the merge.

If we remove it from repo, it still can be generated if someone wants to translate strings, and at the same time, avoid the conflicts on merges.

Patch-set: 2
Reviewer: Pablo Iranzo Gómez <1006779@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0