Update patch set 12
Patch Set 12:

Another thing noticed is that to get plugins to work from tox and from main file:

extension = "core"
try:
    pluginsdir = os.path.join(citellus.citellusdir, 'plugins', extension)
except:
    pluginsdir = os.path.join(citellus.citellus.citellusdir, 'plugins', extension)


def run(plugin):  # do not edit this line
    """
    Executes plugin
    :return: returncode, out, err
    """
    try:
        return citellus.execonshell(filename=plugin['plugin'])
    except:
        return citellus.citellus.execonshell(filename=plugin['plugin'])

same for ansible:

extension = "ansible"                                      
try:                                                       
    pluginsdir = os.path.join(citellus.citellusdir, 'plugins', extension)
except:                                                    
    pluginsdir = os.path.join(citellus.citellus.citellusdir, 'plugins', extension)  

This patch is just to show where the problems are most likely.

First problem: From tox we call docitellus() directly, thus the main() is not run as it is imported and __name__ != __main__ and within main we call initExtensions().

And second problem: because of the location we cant expand citellusdir, in tox its citellus.citellus.citellusdir and in main its citellus.citellusdir.

Patch-set: 12