Update patch set 1
Patch Set 1:

Probably no access to the internet which makes sense, but would like to know where it failed. Was this a unit testing the common functions or a something trying to use it.
If the latter, it would be best to hardcode tests with both(or more) intended used values anyhow

Patch-set: 1
Label: Code-Review=0