Changes to services based on systemctl
is_active and is_enabled were using the same files to check if a service
is active or enabled. For live systems only one command is used but for
non-live systems different sosreport provided systectl commands are used.

Some of this commands can be used to check if a service is enabled or active,
but not both.

Split the array of file that are used for checks for is_active and is_enabled.

Also, added is_running for systemd based services.

Change-Id: I892646bb810fe8cfffcd0fefb58d017255cc15cd