[travis] Automatic Github build and PyPI on tags
this changes the way Travis builds version on Github and PyPI.

Now each time you want to release a version you have to tag the
commit.

For example to release version 1.1.0:

git checkout 1.1.0
git tag -am "Releasing 1.1.0" 1.1.0

Travis will detect this because of on: tags: true and will
initiate build of PyPI package and as well Github release
automatically.

Change-Id: If64b6435e92b63a51c4d2c2850caa78ce355f991